requirements.txt
Pillow==10.0.0
cryptography==41.0.3
numpy==1.24.3